test: move P2.9 prometheus Counter test to integration (U1 policy)
CI lint (check-test-policy) flagged U1-importorskip: the cost/volume Counter test
used pytest.importorskip("prometheus_client") in tests/unit/. Move it to
tests/integration/server/ (importorskip allowed there); the unit file keeps only
the pure-helper tests (parse_iso / discover).
